refactor: update path
