+++ src/server/auth/index.ts
@@ -141,7 +141,7 @@ export const authOptions: NextAuthOptions = {
     })
   ],
   pages: {
-    signIn: '/login'
+    signIn: '/'
   }
 };
